Refactor cache implementations into multiple files
